requirements.txt
asgiref==3.8.1
django==5.2
djangorestframework==3.16.0
joblib==1.5.0
numpy==2.2.5
psycopg==3.2.7
psycopg2==2.9.10
scikit-learn==1.6.1
scipy==1.15.2
sqlparse==0.5.3
threadpoolctl==3.6.0
typing-extensions==4.13.2
tzdata==2025.2
